chore: move makeBrandedAmountPattern; minor improvements
